refactor(use-tooltip): separate the logic in tooltip.tsx into use-tooltip.ts
